requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tqdm
wandb
scipy